Move mail settings to its own component
