+++ src-ui/src/app/app.component.ts
@@ -180,6 +180,14 @@ export class AppComponent implements OnInit, OnDestroy {
           },
         },
         {
+          anchorId: 'tour.mail',
+          content: $localize`Manage e-mail accounts and rules for automatically importing documents.`,
+          route: '/mail',
+          backdropConfig: {
+            offset: 0,
+          },
+        },
+        {
           anchorId: 'tour.file-tasks',
           content: $localize`File Tasks shows you documents that have been consumed, are waiting to be, or may have failed during the process.`,
           route: '/tasks',
@@ -189,7 +197,7 @@ export class AppComponent implements OnInit, OnDestroy {
         },
         {
           anchorId: 'tour.settings',
-          content: $localize`Check out the settings for various tweaks to the web app, toggle settings for saved views or setup e-mail checking.`,
+          content: $localize`Check out the settings for various tweaks to the web app, toggle settings for saved views or manage users.`,
           route: '/settings',
           backdropConfig: {
             offset: 0,
