+++ src-ui/src/app/components/admin/settings/settings.component.ts
@@ -1,65 +1,56 @@
+import { ViewportScroller } from '@angular/common'
 import {
   Component,
+  OnInit,
+  AfterViewInit,
+  OnDestroy,
   Inject,
   LOCALE_ID,
-  OnInit,
-  OnDestroy,
-  AfterViewInit,
 } from '@angular/core'
-import { FormControl, FormGroup } from '@angular/forms'
-import { PaperlessSavedView } from 'src/app/data/paperless-saved-view'
-import { DocumentListViewService } from 'src/app/services/document-list-view.service'
-import { SavedViewService } from 'src/app/services/rest/saved-view.service'
-import {
-  LanguageOption,
-  SettingsService,
-} from 'src/app/services/settings.service'
-import { Toast, ToastService } from 'src/app/services/toast.service'
-import { dirtyCheck, DirtyComponent } from '@ngneat/dirty-check-forms'
-import {
-  Observable,
-  Subscription,
-  BehaviorSubject,
-  first,
-  tap,
-  takeUntil,
-  Subject,
-} from 'rxjs'
-import { SETTINGS_KEYS } from 'src/app/data/paperless-uisettings'
+import { FormGroup, FormControl } from '@angular/forms'
 import { ActivatedRoute, Router } from '@angular/router'
-import { ViewportScroller } from '@angular/common'
-import { TourService } from 'ngx-ui-tour-ng-bootstrap'
-import { ComponentWithPermissions } from '../../with-permissions/with-permissions.component'
 import { NgbModal, NgbNavChangeEvent } from '@ng-bootstrap/ng-bootstrap'
-import { UserService } from 'src/app/services/rest/user.service'
-import { GroupService } from 'src/app/services/rest/group.service'
-import { PaperlessUser } from 'src/app/data/paperless-user'
-import { PaperlessGroup } from 'src/app/data/paperless-group'
-import { UserEditDialogComponent } from '../../common/edit-dialog/user-edit-dialog/user-edit-dialog.component'
-import { ConfirmDialogComponent } from '../../common/confirm-dialog/confirm-dialog.component'
-import { GroupEditDialogComponent } from '../../common/edit-dialog/group-edit-dialog/group-edit-dialog.component'
-import { PaperlessMailAccount } from 'src/app/data/paperless-mail-account'
-import { PaperlessMailRule } from 'src/app/data/paperless-mail-rule'
-import { MailAccountService } from 'src/app/services/rest/mail-account.service'
-import { MailRuleService } from 'src/app/services/rest/mail-rule.service'
-import { MailAccountEditDialogComponent } from '../../common/edit-dialog/mail-account-edit-dialog/mail-account-edit-dialog.component'
-import { MailRuleEditDialogComponent } from '../../common/edit-dialog/mail-rule-edit-dialog/mail-rule-edit-dialog.component'
-import { EditDialogMode } from '../../common/edit-dialog/edit-dialog.component'
-import { ObjectWithPermissions } from 'src/app/data/object-with-permissions'
+import { DirtyComponent, dirtyCheck } from '@ngneat/dirty-check-forms'
+import { TourService } from 'ngx-ui-tour-ng-bootstrap'
 import {
+  BehaviorSubject,
+  Subscription,
+  Observable,
+  Subject,
+  first,
+  takeUntil,
+  tap,
+} from 'rxjs'
+import { ObjectWithPermissions } from 'src/app/data/object-with-permissions'
+import { PaperlessGroup } from 'src/app/data/paperless-group'
+import { PaperlessSavedView } from 'src/app/data/paperless-saved-view'
+import { SETTINGS_KEYS } from 'src/app/data/paperless-uisettings'
+import { PaperlessUser } from 'src/app/data/paperless-user'
+import { DocumentListViewService } from 'src/app/services/document-list-view.service'
+import {
+  PermissionsService,
   PermissionAction,
   PermissionType,
-  PermissionsService,
 } from 'src/app/services/permissions.service'
-import { PermissionsDialogComponent } from '../../common/permissions-dialog/permissions-dialog.component'
-import { AbstractPaperlessService } from 'src/app/services/rest/abstract-paperless-service'
+import { GroupService } from 'src/app/services/rest/group.service'
+import { SavedViewService } from 'src/app/services/rest/saved-view.service'
+import { UserService } from 'src/app/services/rest/user.service'
+import {
+  SettingsService,
+  LanguageOption,
+} from 'src/app/services/settings.service'
+import { ToastService, Toast } from 'src/app/services/toast.service'
+import { ConfirmDialogComponent } from '../../common/confirm-dialog/confirm-dialog.component'
+import { EditDialogMode } from '../../common/edit-dialog/edit-dialog.component'
+import { GroupEditDialogComponent } from '../../common/edit-dialog/group-edit-dialog/group-edit-dialog.component'
+import { UserEditDialogComponent } from '../../common/edit-dialog/user-edit-dialog/user-edit-dialog.component'
+import { ComponentWithPermissions } from '../../with-permissions/with-permissions.component'
 
 enum SettingsNavIDs {
   General = 1,
   Notifications = 2,
   SavedViews = 3,
-  Mail = 4,
-  UsersGroups = 5,
+  UsersGroups = 4,
 }
 
 @Component({
@@ -77,9 +68,6 @@ export class SettingsComponent
   savedViewGroup = new FormGroup({})
   usersGroup = new FormGroup({})
   groupsGroup = new FormGroup({})
-
-  mailAccountGroup = new FormGroup({})
-  mailRuleGroup = new FormGroup({})
 
   settingsForm = new FormGroup({
     bulkEditConfirmationDialogs: new FormControl(null),
@@ -111,15 +99,9 @@ export class SettingsComponent
 
     savedViewsWarnOnUnsavedChange: new FormControl(null),
     savedViews: this.savedViewGroup,
-
-    mailAccounts: this.mailAccountGroup,
-    mailRules: this.mailRuleGroup,
   })
 
   savedViews: PaperlessSavedView[]
-
-  mailAccounts: PaperlessMailAccount[]
-  mailRules: PaperlessMailRule[]
 
   store: BehaviorSubject<any>
   storeSub: Subscription
@@ -141,8 +123,6 @@ export class SettingsComponent
 
   constructor(
     public savedViewService: SavedViewService,
-    public mailAccountService: MailAccountService,
-    public mailRuleService: MailRuleService,
     private documentListViewService: DocumentListViewService,
     private toastService: ToastService,
     private settings: SettingsService,
@@ -165,25 +145,66 @@ export class SettingsComponent
   ngOnInit() {
     this.initialize()
 
+    if (
+      this.permissionsService.currentUserCan(
+        PermissionAction.View,
+        PermissionType.User
+      )
+    ) {
+      this.usersService
+        .listAll()
+        .pipe(first())
+        .subscribe({
+          next: (r) => {
+            this.users = r.results
+            this.initialize(false)
+          },
+          error: (e) => {
+            this.toastService.showError($localize`Error retrieving users`, e)
+          },
+        })
+    }
+
+    if (
+      this.permissionsService.currentUserCan(
+        PermissionAction.View,
+        PermissionType.Group
+      )
+    ) {
+      this.groupsService
+        .listAll()
+        .pipe(first())
+        .subscribe({
+          next: (r) => {
+            this.groups = r.results
+            this.initialize(false)
+          },
+          error: (e) => {
+            this.toastService.showError($localize`Error retrieving groups`, e)
+          },
+        })
+    }
+
+    if (
+      this.permissionsService.currentUserCan(
+        PermissionAction.View,
+        PermissionType.SavedView
+      )
+    ) {
+      this.savedViewService.listAll().subscribe((r) => {
+        this.savedViews = r.results
+        this.initialize(false)
+      })
+    }
+
     this.activatedRoute.paramMap.subscribe((paramMap) => {
       const section = paramMap.get('section')
-      if (section === null) {
-        if (
-          this.permissionsService.currentUserCan(
-            PermissionAction.View,
-            PermissionType.User
-          )
-        ) {
-          this.getUsers()
-        }
-      }
       if (section) {
         const navIDKey: string = Object.keys(SettingsNavIDs).find(
           (navID) => navID.toLowerCase() == section
         )
         if (navIDKey) {
           this.activeNavID = SettingsNavIDs[navIDKey]
-          this.maybeInitializeTab(this.activeNavID)
         }
       }
     })
@@ -256,13 +277,10 @@ export class SettingsComponent
       usersGroup: {},
       groupsGroup: {},
       savedViews: {},
-      mailAccounts: {},
-      mailRules: {},
     }
   }
 
   onNavChange(navChangeEvent: NgbNavChangeEvent) {
-    this.maybeInitializeTab(navChangeEvent.nextId)
     const [foundNavIDkey] = Object.entries(SettingsNavIDs).find(
       ([, navIDValue]) => navIDValue == navChangeEvent.nextId
     )
@@ -277,83 +295,6 @@ export class SettingsComponent
             this.initialize()
           }
         })
-  }
-
-  // Load tab contents 'on demand', either on mouseover or focusin (i.e. before click) or called from nav change event
-  maybeInitializeTab(navID: number): void {
-    if (navID == SettingsNavIDs.SavedViews && !this.savedViews) {
-      this.savedViewService.listAll().subscribe((r) => {
-        this.savedViews = r.results
-        this.initialize(false)
-      })
-    } else if (
-      (navID == SettingsNavIDs.UsersGroups ||
-        navID == SettingsNavIDs.General) &&
-      (!this.users || !this.groups)
-    ) {
-      if (!this.users) this.getUsers()
-      this.groupsService
-        .listAll()
-        .pipe(first())
-        .subscribe({
-          next: (r) => {
-            this.groups = r.results
-            this.initialize(false)
-          },
-          error: (e) => {
-            this.toastService.showError($localize`Error retrieving groups`, e)
-          },
-        })
-    } else if (
-      navID == SettingsNavIDs.Mail &&
-      (!this.mailAccounts || !this.mailRules)
-    ) {
-      this.mailAccountService
-        .listAll(null, null, { full_perms: true })
-        .pipe(first())
-        .subscribe({
-          next: (r) => {
-            this.mailAccounts = r.results
-
-            this.mailRuleService
-              .listAll(null, null, { full_perms: true })
-              .pipe(first())
-              .subscribe({
-                next: (r) => {
-                  this.mailRules = r.results
-                  this.initialize(false)
-                },
-                error: (e) => {
-                  this.toastService.showError(
-                    $localize`Error retrieving mail rules`,
-                    e
-                  )
-                },
-              })
-          },
-          error: (e) => {
-            this.toastService.showError(
-              $localize`Error retrieving mail accounts`,
-              e
-            )
-          },
-        })
-    }
-  }
-
-  private getUsers() {
-    this.usersService
-      .listAll()
-      .pipe(first())
-      .subscribe({
-        next: (r) => {
-          this.users = r.results
-          this.initialize(false)
-        },
-        error: (e) => {
-          this.toastService.showError($localize`Error retrieving users`, e)
-        },
-      })
   }
 
   initialize(resetSettings: boolean = true) {
@@ -414,6 +355,7 @@ export class SettingsComponent
         )
       }
     }
+
     if (this.groups) {
       this.emptyGroup(this.groupsGroup)
       for (let group of this.groups) {
@@ -428,81 +370,6 @@ export class SettingsComponent
             id: new FormControl(null),
             name: new FormControl(null),
             permissions: new FormControl(null),
-          })
-        )
-      }
-    }
-
-    if (this.mailAccounts && this.mailRules) {
-      this.emptyGroup(this.mailAccountGroup)
-      this.emptyGroup(this.mailRuleGroup)
-
-      for (let account of this.mailAccounts) {
-        storeData.mailAccounts[account.id.toString()] = {
-          id: account.id,
-          name: account.name,
-          imap_server: account.imap_server,
-          imap_port: account.imap_port,
-          imap_security: account.imap_security,
-          username: account.username,
-          password: account.password,
-          character_set: account.character_set,
-        }
-        this.mailAccountGroup.addControl(
-          account.id.toString(),
-          new FormGroup({
-            id: new FormControl(null),
-            name: new FormControl(null),
-            imap_server: new FormControl(null),
-            imap_port: new FormControl(null),
-            imap_security: new FormControl(null),
-            username: new FormControl(null),
-            password: new FormControl(null),
-            character_set: new FormControl(null),
-          })
-        )
-      }
-
-      for (let rule of this.mailRules) {
-        storeData.mailRules[rule.id.toString()] = {
-          name: rule.name,
-          account: rule.account,
-          folder: rule.folder,
-          filter_from: rule.filter_from,
-          filter_to: rule.filter_to,
-          filter_subject: rule.filter_subject,
-          filter_body: rule.filter_body,
-          filter_attachment_filename: rule.filter_attachment_filename,
-          maximum_age: rule.maximum_age,
-          attachment_type: rule.attachment_type,
-          action: rule.action,
-          action_parameter: rule.action_parameter,
-          assign_title_from: rule.assign_title_from,
-          assign_tags: rule.assign_tags,
-          assign_document_type: rule.assign_document_type,
-          assign_correspondent_from: rule.assign_correspondent_from,
-          assign_correspondent: rule.assign_correspondent,
-        }
-        this.mailRuleGroup.addControl(
-          rule.id.toString(),
-          new FormGroup({
-            name: new FormControl(null),
-            account: new FormControl(null),
-            folder: new FormControl(null),
-            filter_from: new FormControl(null),
-            filter_to: new FormControl(null),
-            filter_subject: new FormControl(null),
-            filter_body: new FormControl(null),
-            filter_attachment_filename: new FormControl(null),
-            maximum_age: new FormControl(null),
-            attachment_type: new FormControl(null),
-            action: new FormControl(null),
-            action_parameter: new FormControl(null),
-            assign_title_from: new FormControl(null),
-            assign_tags: new FormControl(null),
-            assign_document_type: new FormControl(null),
-            assign_correspondent_from: new FormControl(null),
-            assign_correspondent: new FormControl(null),
           })
         )
       }
@@ -722,17 +589,17 @@ export class SettingsComponent
       x.push(this.savedViewGroup.value[id])
     }
     if (x.length > 0) {
-      this.savedViewService.patchMany(x).subscribe(
-        (s) => {
+      this.savedViewService.patchMany(x).subscribe({
+        next: () => {
           this.saveLocalSettings()
         },
-        (error) => {
+        error: (error) => {
           this.toastService.showError(
             $localize`Error while storing settings on server.`,
             error
           )
-        }
-      )
+        },
+      })
     } else {
       this.saveLocalSettings()
     }
@@ -873,153 +740,5 @@ export class SettingsComponent
 
   getGroupName(id: number): string {
     return this.groups?.find((g) => g.id === id)?.name ?? ''
-  }
-
-  editMailAccount(account: PaperlessMailAccount) {
-    const modal = this.modalService.open(MailAccountEditDialogComponent, {
-      backdrop: 'static',
-      size: 'xl',
-    })
-    modal.componentInstance.dialogMode = account
-      ? EditDialogMode.EDIT
-      : EditDialogMode.CREATE
-    modal.componentInstance.object = account
-    modal.componentInstance.succeeded
-      .pipe(takeUntil(this.unsubscribeNotifier))
-      .subscribe((newMailAccount) => {
-        this.toastService.showInfo(
-          $localize`Saved account "${newMailAccount.name}".`
-        )
-        this.mailAccountService.clearCache()
-        this.mailAccountService
-          .listAll(null, null, { full_perms: true })
-          .subscribe((r) => {
-            this.mailAccounts = r.results
-            this.initialize()
-          })
-      })
-    modal.componentInstance.failed
-      .pipe(takeUntil(this.unsubscribeNotifier))
-      .subscribe((e) => {
-        this.toastService.showError($localize`Error saving account.`, e)
-      })
-  }
-
-  deleteMailAccount(account: PaperlessMailAccount) {
-    const modal = this.modalService.open(ConfirmDialogComponent, {
-      backdrop: 'static',
-    })
-    modal.componentInstance.title = $localize`Confirm delete mail account`
-    modal.componentInstance.messageBold = $localize`This operation will permanently delete this mail account.`
-    modal.componentInstance.message = $localize`This operation cannot be undone.`
-    modal.componentInstance.btnClass = 'btn-danger'
-    modal.componentInstance.btnCaption = $localize`Proceed`
-    modal.componentInstance.confirmClicked.subscribe(() => {
-      modal.componentInstance.buttonsEnabled = false
-      this.mailAccountService.delete(account).subscribe({
-        next: () => {
-          modal.close()
-          this.toastService.showInfo($localize`Deleted mail account`)
-          this.mailAccountService.clearCache()
-          this.mailAccountService
-            .listAll(null, null, { full_perms: true })
-            .subscribe((r) => {
-              this.mailAccounts = r.results
-              this.initialize(true)
-            })
-        },
-        error: (e) => {
-          this.toastService.showError(
-            $localize`Error deleting mail account.`,
-            e
-          )
-        },
-      })
-    })
-  }
-
-  editMailRule(rule: PaperlessMailRule) {
-    const modal = this.modalService.open(MailRuleEditDialogComponent, {
-      backdrop: 'static',
-      size: 'xl',
-    })
-    modal.componentInstance.dialogMode = rule
-      ? EditDialogMode.EDIT
-      : EditDialogMode.CREATE
-    modal.componentInstance.object = rule
-    modal.componentInstance.succeeded
-      .pipe(takeUntil(this.unsubscribeNotifier))
-      .subscribe((newMailRule) => {
-        this.toastService.showInfo($localize`Saved rule "${newMailRule.name}".`)
-        this.mailRuleService.clearCache()
-        this.mailRuleService
-          .listAll(null, null, { full_perms: true })
-          .subscribe((r) => {
-            this.mailRules = r.results
-
-            this.initialize(true)
-          })
-      })
-    modal.componentInstance.failed
-      .pipe(takeUntil(this.unsubscribeNotifier))
-      .subscribe((e) => {
-        this.toastService.showError($localize`Error saving rule.`, e)
-      })
-  }
-
-  deleteMailRule(rule: PaperlessMailRule) {
-    const modal = this.modalService.open(ConfirmDialogComponent, {
-      backdrop: 'static',
-    })
-    modal.componentInstance.title = $localize`Confirm delete mail rule`
-    modal.componentInstance.messageBold = $localize`This operation will permanently delete this mail rule.`
-    modal.componentInstance.message = $localize`This operation cannot be undone.`
-    modal.componentInstance.btnClass = 'btn-danger'
-    modal.componentInstance.btnCaption = $localize`Proceed`
-    modal.componentInstance.confirmClicked.subscribe(() => {
-      modal.componentInstance.buttonsEnabled = false
-      this.mailRuleService.delete(rule).subscribe({
-        next: () => {
-          modal.close()
-          this.toastService.showInfo($localize`Deleted mail rule`)
-          this.mailRuleService.clearCache()
-          this.mailRuleService
-            .listAll(null, null, { full_perms: true })
-            .subscribe((r) => {
-              this.mailRules = r.results
-              this.initialize(true)
-            })
-        },
-        error: (e) => {
-          this.toastService.showError($localize`Error deleting mail rule.`, e)
-        },
-      })
-    })
-  }
-
-  editPermissions(object: PaperlessMailRule | PaperlessMailAccount) {
-    const modal = this.modalService.open(PermissionsDialogComponent, {
-      backdrop: 'static',
-    })
-    const dialog: PermissionsDialogComponent =
-      modal.componentInstance as PermissionsDialogComponent
-    dialog.object = object
-    modal.componentInstance.confirmClicked.subscribe((permissions) => {
-      modal.componentInstance.buttonsEnabled = false
-      const service: AbstractPaperlessService<
-        PaperlessMailRule | PaperlessMailAccount
-      > = 'account' in object ? this.mailRuleService : this.mailAccountService
-      object.owner = permissions['owner']
-      object['set_permissions'] = permissions['set_permissions']
-      service.patch(object).subscribe({
-        next: () => {
-          this.toastService.showInfo($localize`Permissions updated`)
-          modal.close()
-        },
-        error: (e) => {
-          this.toastService.showError($localize`Error updating permissions`, e)
-        },
-      })
-    })
   }
 }
